requirements.txt
# 核心Django依赖
Django==4.2.7
asgiref==3.9.1
sqlparse==0.5.3
pytz==2024.1

# REST Framework相关
djangorestframework==3.16.0
djangorestframework_simplejwt==5.3.0
PyJWT==2.9.0
rest_framework_mongoengine==3.5.0
# 过滤器支持
django-filter==23.5

# API文档
drf-spectacular==0.28.0
Pygments==2.19.2
uritemplate==4.2.0

# MongoDB相关
mongoengine==0.29.1
pymongo==3.11.4

# 类型提示增强
typing_extensions==4.14.1

# 以下是之前实现的Celery相关功能所需依赖
celery==5.4.0
redis==6.2.0
django-redis==6.0.0

# CORS支持
django-cors-headers==4.7.0

# LLM处理相关依赖
langchain==0.2.0
langchain_openai==0.1.0

# 其他可能需要的实用工具包
python-dotenv==1.1.1
inflection==0.5.1
packaging==25.0
six==1.16.0
python-dateutil==2.9.0.post0

# 开发和部署工具
wheel==0.45.1
setuptools==78.1.1
